enterprise-data-pipeline/requirements.txt
# Core Dependencies
requests==2.31.0
pyyaml==6.0.1
tenacity==8.2.3
python-dotenv==1.0.0

# Databricks/Spark
pyspark==3.5.0
databricks-sdk==0.18.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Snowflake
snowflake-connector-python[pandas]==3.6.0
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Logging & Monitoring
python-json-logger==2.0.7

# Code Quality (optional for development)
black==23.12.1
flake8==7.0.0
mypy==1.7.1
types-requests==2.31.0.10
types-PyYAML==6.0.12.12

# Documentation (optional)
mkdocs==1.5.3
mkdocs-material==9.5.3
